Grid movement

+++ src/main/java/components/UserCommitmentComponent.java
@@ -6,7 +6,7 @@ import com.badlogic.ashley.core.Component;
  * Created by Thagus on 18/10/16.
  */
 public class UserCommitmentComponent implements Component {
-    public int xCommitment, yCommitment;
+    public float xCommitment, yCommitment;
     public int xDestination, yDestination;
     public CityBlockComponent.BlockType destination;
 }
